优化诊断禁用
