+++ crates/emmylua_ls/src/handlers/code_actions/actions/mod.rs
@@ -1,1 +1,3 @@
+mod build_disable_code;
 
+pub use build_disable_code::*;
